chore(setup): Using QuestOwl as a template (read des.)
QuestOwl is a repo at the time of this commit. This repo was essentially the product for the FBLA National Leadership Conference 2025 in Anaheim, California. However, it is pretty buggy, and kind of weird. I might either finish it, leave it to sit here, or delete it. It is a great starting block to our repository!

This new project might be successful, and there are currently 2 people working on this project for FBLA (including me)!